Move synthetic-manifest generator to `metricflow-semantics`.
