+++ metricflow-semantics/metricflow_semantics/test_helpers/synthetic_manifest/measure_generator.py
@@ -1,6 +1,8 @@
 from __future__ import annotations
 
-from tests_metricflow.performance.synthetic_manifest_parameter_set import SyntheticManifestParameterSet
+from metricflow_semantics.test_helpers.synthetic_manifest.synthetic_manifest_parameter_set import (
+    SyntheticManifestParameterSet,
+)
 
 
 class MeasureGenerator:
