+++ metricflow-semantics/metricflow_semantics/test_helpers/synthetic_manifest/measure_semantic_model_generator.py
@@ -9,8 +9,10 @@ from dbt_semantic_interfaces.implementations.node_relation import PydanticNodeRe
 from dbt_semantic_interfaces.implementations.semantic_model import PydanticSemanticModel
 from dbt_semantic_interfaces.type_enums import AggregationType, DimensionType, EntityType, TimeGranularity
 
-from tests_metricflow.performance.measure_generator import MeasureGenerator
-from tests_metricflow.performance.synthetic_manifest_parameter_set import SyntheticManifestParameterSet
+from metricflow_semantics.test_helpers.synthetic_manifest.measure_generator import MeasureGenerator
+from metricflow_semantics.test_helpers.synthetic_manifest.synthetic_manifest_parameter_set import (
+    SyntheticManifestParameterSet,
+)
 
 
 class MeasureSemanticModelGenerator:
